Break down home page into smaller subcomponents
