+++ src/app/_components/error.tsx
@@ -1,0 +1,3 @@
+export const PageError = ({ message }: { message: string }) => {
+  return <div className="p-8 text-red-500">Error: {message}</div>;
+};
