+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
-import { ReactQueryProvider } from '@/providers/ReactQueryProvider';
 import './globals.css';
+import { ReactQueryProvider } from '@/providers/reactQueryProvider';
 const inter = Inter({ subsets: ['latin'] });
 
 export const metadata: Metadata = {
